[PowerRename] Refactor common regex tests and introduce advanced counter tests
